ASoC: msm8916-wcd-analog: Add jack detection using GPIO
Samsung decided to implement entirely custom jack detection using
an extra IC for some reason. Because of that, the jack detection
in PM8916 does not work as-is. (It always detects that headphones
are plugged in...)

However, the headset/button detection is still working fine through
PM8916. To avoid implementing custom Samsung stuff, override the
jack detection to use the provided GPIO, but keep everything else.

Signed-off-by: Stephan Gerhold <stephan@gerhold.net>